ContikiMoteType: use arena for scope handling
Add added/removed methods to the MoteType
interface, and call that from the simulation
when the motetype is added/removed.

Inside ContikiMoteType, replace the current
class-loader-tweaking with an arena that
is closed when the motetype is removed from
the simulation. This will close the shared
library as well.

This reduces the runtime for 07-simulation-base
from 1m 8s to 50s.
